Allow the user to change the models' display names
Previously, the model's ID would be used as the display name. Now, the
user can change it to a custom name. This is done very similarly to the
`API_MODELS` field: the user provides a comma-separated list of display
names in the `API_MODEL_NAMES` field.

Nominally, the user should provide the same number of display names as
the number of models. If they provide fewer, the display names will be
extended with `None`. If they provide more, the extra ones will be
ignored. Also note that the display names are copied verbatim - we don't
prefix them with `reasoning/`, for example.